Fix local reservation flag being cleared by trap entry as well as exit.
On closer re-reading of the priv spec, it only gives permission for xRET
to clear the reservation. This is a bug, albeit a minor one, since the
intent is obviously just 'taking a trap clears your reservation.'

This continues to not affect debug mode entry and exit, so you can step
through LR/SC sequences.