Fix: dphase fault on final load of a cm.popret with at least two loads samples wrong exception pc.
A special case was added to the CIR-lock logic when Zcmp was added which
allows the PC to update even though decode is stalled, because the jr
is actually the penultimate uop so that the final `add sp` can execute in the fetch bubble.

This PC update is premature if the jr uop takes multiple cycles, as is
always the case when a bus fault response is returned to the immediately
preceding load.